+++ src/store/mod.rs
@@ -124,6 +124,7 @@ pub mod tests {
                 "user1" | "user2" | "user3" => Some(User {
                     name: key.to_string(),
                     password: Password::Plain(key.to_string()),
+                    allowed_scopes: Default::default(),
                     attributes: HashMap::new(),
                 }),
                 _ => None,